refactor: interface 디렉터리 명 types로 변경 interface도 있고 type도 있고 섞여 있는데다가 의미상 types로 하는게 맞는 것 같아서 변경
